requirements_free.txt
streamlit
python-dotenv
groq
langchain
sentence-transformers
pymupdf
docx2txt
faiss-cpu